add 6 bits for algorithm in the TrackBase::Status class,  the enum is kept in TrackBase similar to CMSSW case; set it in the WriteMemoryFile.cc
